# api/swagger/swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Assets Service",
    "version": "1.0"
  },
  "host": "localhost:3005",
  "basePath": "/",
  "paths": {
    "/api/asset/{name}": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "summary": "Get users's asset by name",
        "operationId": "LoadAsset",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AssetName",
            "description": "asset name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/StatusResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "summary": "Delete users's asset by name",
        "operationId": "DeleteAsset",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AssetName",
            "description": "asset name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/StatusResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/assets": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "summary": "Get users's assets list",
        "operationId": "LoadAssetsList",
        "responses": {
          "200": {
            "$ref": "#/responses/AssetsListResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/auth": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Authenicates user",
        "operationId": "Authenicate",
        "parameters": [
          {
            "example": "{\"login\": \"alice\", \"password\": \"secret\"}",
            "description": "credentials of the user",
            "name": "Credentials",
            "in": "body",
            "required": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TokenResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/upload-asset/{name}": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "any"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "summary": "Store asset",
        "operationId": "StoreAsset",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AssetName",
            "description": "asset name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "type": "file",
            "x-go-name": "Data",
            "description": "asset data",
            "name": "data",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/StatusResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Creates user",
        "operationId": "CreateUser",
        "parameters": [
          {
            "example": "{\"login\": \"alice\", \"password\": \"secret\"}",
            "description": "credentials of the user",
            "name": "Credentials",
            "in": "body",
            "required": true,
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/StatusResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/health": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "application"
        ],
        "summary": "Cumulative information about the readiness and performance of the service",
        "operationId": "health",
        "responses": {
          "200": {
            "$ref": "#/responses/AppInfoResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "responses": {
    "AppInfoResponse": {
      "description": "Cumulative information about the readiness and performance of the service",
      "headers": {
        "state": {
          "type": "string",
          "example": "\"running\"",
          "description": "state"
        },
        "version": {
          "type": "string",
          "example": "\"1.0\"",
          "description": "version"
        }
      }
    },
    "AssetsListResponse": {
      "description": "Success get assets list response",
      "headers": {
        "assets": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": "\"[file1.txt, file2.txt, file3.txt]\"",
          "description": "assets"
        }
      }
    },
    "ErrorResponse": {
      "description": "Common error response",
      "headers": {
        "error": {
          "type": "string",
          "example": "\"Internal server error\"",
          "description": "error"
        }
      }
    },
    "StatusResponse": {
      "description": "Common success response",
      "headers": {
        "status": {
          "type": "string",
          "example": "\"ok\"",
          "description": "status"
        }
      }
    },
    "TokenResponse": {
      "description": "Success authenication response",
      "headers": {
        "token": {
          "type": "string",
          "example": "\"b5a302e740d0d84bbdc2254c97f1427b\"",
          "description": "token"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}